Remove --with-merge option
This option was introduced to avoid github conflict issues by creating
an empty merge commit. But it turned out that this commit creates
issues for future rebases.

After some consideration we decided that it's better to resolve such
commits manually, and bot's help is not required. After this we can
remove --with-merge option.