Fix PR open/update depending on push
When destination needs updating but rebase branch is up to date.
If there was no PR it would not be created and if there was PR but it
had incorrect title it would not be updated.

This PR makes it always check for PR if the dest branch is missing
commit and always computes PR title and updates it when required